% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_ngl.R
\name{color_key_match}
\alias{color_key_match}
\title{Internal utility function for matching available NGL colorSchemes}
\usage{
color_key_match(colorScheme)
}
\description{
This function is for internal use only. No guarantee is provided for its
stability across package versions.
}
\keyword{internal}
