% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_ngl.R
\name{pca2string}
\alias{pca2string}
\title{Convert a bio3d PCA or NMA object to a character vector for NGLVieweR input}
\usage{
pca2string(pc, ...)
}
\arguments{
\item{pc}{the results of principal component analysis or normal mode analysis as obtained with \code{pca()} or \code{nma()} and friends.}

\item{...}{Extra arguments passed to \code{mktrj()}.}
}
\value{
a character vector of structure data.
}
\description{
Convert a bio3d PCA or NMA object to a character vector for NGLVieweR input
}
\examples{
 #pdb <- bio3d::read.pdb("6s36")
 #n <- bio3d::nma(pdb)
 #NGLVieweR::NGLVieweR( pca2string(n), format="pdb") |>
 #   NGLVieweR::addRepresentation("cartoon")

 # Or more simpley...
 # view.pdb(n) # Does not work yet!!!
 # view.pdb(n, colorScheme = "residueindex")

}
\seealso{
\code{view.pdb()}, \code{pdb2string()}, \code{NGLVieweR::NGLVieweR()}, \code{bio3d::nma()}, \code{bio3d::pca()}, \code{bio3d::mktrj()}.
}
\author{
Barry Grant, \email{bjgrant@ucsd.edu}
}
