% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_ngl.R
\name{pdb2ngl}
\alias{pdb2ngl}
\title{Convert a bio3d PDB object for NGLVieweR}
\usage{
pdb2ngl(pdb)
}
\arguments{
\item{pdb}{a bio3d pdb object as obtained from \code{read.pdb()}.}
}
\value{
a single element character vector with return characters as required by \code{NGLVieweR::NGLVieweR()} function with \code{format=pdb} option.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\details{
Please use \code{pdb2string()} instead.
Function to take a bio3d structure and use in the NGLVieweR package.
}
\examples{
pdb <- bio3d::read.pdb("5p21")
NGLVieweR::NGLVieweR(pdb2ngl(pdb), format="pdb") |>
  NGLVieweR::addRepresentation("cartoon")

}
\seealso{
\code{view.pdb()} which uses this function internally, \code{NGLVieweR::NGLVieweR()}, \code{bio3d::read.pdb()}.
}
\author{
Barry Grant, \email{bjgrant@ucsd.edu}
}
\keyword{internal}
