% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_ngl.R
\name{pdb2string}
\alias{pdb2string}
\title{Convert a PDB Object to a Character String}
\usage{
pdb2string(pdb, ...)
}
\arguments{
\item{pdb}{a bio3d pdb object as obtained from \code{read.pdb()}.}

\item{...}{Extra arguments that are passed to \code{write.pdb()}.}
}
\value{
a single element character vector with return characters as required
by \code{NGLVieweR::NGLVieweR()} function with \code{format=pdb} option. Note that
no intermediate files are produced.
}
\description{
Function to take a \strong{bio3d} \code{pdb} structure object and convert it to a single
element character vector that can then be used as input for the \strong{NGLVieweR}
and \strong{r3dmol} packages.
}
\examples{
pdb <- bio3d::read.pdb("5p21")

NGLVieweR::NGLVieweR(pdb2string(pdb), format="pdb") |>
  NGLVieweR::addRepresentation("cartoon")

# Or more simply
view.pdb(pdb)

}
\seealso{
\code{view.pdb()} which uses this function internally, \code{NGLVieweR::NGLVieweR()}, \code{bio3d::read.pdb()}.
}
\author{
Barry Grant, \email{bjgrant@ucsd.edu}
}
