% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_ngl.R
\name{style_key_match}
\alias{style_key_match}
\title{Internal utility function for matching available NGL representations}
\usage{
style_key_match(representation)
}
\description{
This function is for internal use only. No guarantee is provided for its
stability across package versions.
}
\keyword{internal}
