% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_ngl.R
\name{view.pca}
\alias{view.pca}
\alias{view.nma}
\title{Quick molecular viewing of bio3d PCA results}
\usage{
view.pca(
  pc,
  colorScheme = "residueindex",
  representation = "cartoon",
  backgroundColor = "white",
  ...
)

view.nma(
  pc,
  colorScheme = "residueindex",
  representation = "cartoon",
  backgroundColor = "white",
  ...
)
}
\arguments{
\item{pc}{an object of class "pca" as obtained with function \code{pca.xyz} or \code{pca}.}

\item{colorScheme}{keyword based coloring used only if \code{cols} input is NULL. Possible values include "residueindex", "modelindex", "sstruc", "bfactor", "chainid", "chainindex", "atomindex", and "occupancy".}

\item{representation}{the representation style, useful values are "line", "tube", "cartoon", "trace", "backbone", and "ball+stick".}

\item{backgroundColor}{a single element color vector that set the display area background color.}

\item{...}{additional arguments passed to and from functions (e.g. to function \code{write.pdb}).}
}
\value{
an \strong{NGLVieweR} display object that can be displayed or further added to using \code{NGLVieweR::addRepresentation()} and friends.
}
\description{
Quick molecular viewing of bio3d PCA results
}
\section{Functions}{
\itemize{
\item \code{view.nma()}: Quick viewing of NMA modes

}}
\examples{
#pdb <- bio3d::read.pdb("6s36")
#n <- bio3d::nma(pdb)
#view.nma(n)
#view.nma(n, colorScheme = "model")
}
\author{
Barry Grant, \email{bjgrant@ucsd.edu}
}
