% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_ngl.R
\name{view.pdbs2}
\alias{view.pdbs2}
\title{Quick interactive PDBS object viewing using NGLVieweR}
\usage{
view.pdbs2(
  pdbs,
  colors = NULL,
  representation = "line",
  backgroundColor = "white"
)
}
\arguments{
\item{pdbs}{a multi-structure \code{pdbs} object as obtained from \code{pdbaln()}, \code{read.fasta.pdb()}, etc.}

\item{colors}{a vector of colors for each structure. If NULL then the output of \code{vmd.colors()} is used.}

\item{representation}{the representation style, usefull values are lines, tube and cartoon.}

\item{backgroundColor}{set the display area background color.}
}
\value{
an \strong{NGLVieweR} display object that can be displayed or further added to using \code{NGLVieweR::addRepresentation()} and friends.
}
\description{
Quick interactive PDBS object viewing using NGLVieweR
}
\examples{
 # pth <- "~/Desktop/courses/BIMM143/class10/pdbs/split_chain/"
 # files <- list.files(path=pth, full.names = T)
 # pdbs <- pdbaln(files, fit=T, exefile="msa")

 view.pdbs(pdbs, representation = "cartoon")
 view.pdbs(pdbs, colors = c("red","blue") )

}
\seealso{
\code{view.pdb()}, \code{pdb2ngl()}, \code{NGLVieweR::NGLVieweR()}, \code{bio3d::read.pdb()}
}
\author{
Barry Grant, \email{bjgrant@ucsd.edu}
}
