% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_ngl.R
\name{view2png}
\alias{view2png}
\title{Render any htmlwidget to a PNG file}
\usage{
view2png(
  widget,
  file = "snapshot.png",
  width = 800,
  height = 600,
  delay = 0.2,
  autocrop = FALSE
)
}
\arguments{
\item{file}{file name of output image. Should end with an image file type
(‘.png’, ‘.jpg’, ‘.jpeg’, or ‘.webp’) or ‘.pdf’.}

\item{width}{Viewport width}

\item{height}{Viewport height}

\item{delay}{Time to wait before taking screenshot, in seconds. Sometimes a
longer delay is needed for all assets to display properly.}

\item{autocrop}{logical, if TRUE the \code{magic} package is used to trim/remove
white space around the resulting image}
}
\value{
returns image file name
}
\description{
Render any htmlwidget to a PNG file
}
\examples{
#ras <- read.pdb("5p21")
#v <- view.pdb(ras)
#view2png(v)
}
